requirements.txt
Flask==3.0.3
biopython==1.84
pandas==2.2.3
fpdf==1.7.2
matplotlib== 3.9.3
seaborn== 0.13.2

Werkzeug==3.1.3 
